// types/types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};
// ...
export type Game = {
  __typename?: 'Game';
  awayTeam?: Maybe<Team>;
  createdAt?: Maybe<Scalars['String']>;
  homeTeam?: Maybe<Team>;
  id: Scalars['ID'];
  picks?: Maybe<Array<Pick>>;
  picksCount?: Maybe<Scalars['Int']>;
  season?: Maybe<Scalars['String']>;
  slug?: Maybe<Scalars['String']>;
  spread?: Maybe<Scalars['Float']>;
  week?: Maybe<Week>;
  winner?: Maybe<Team>;
};
// ...
export type PasswordState = {
  __typename?: 'PasswordState';
  isSet: Scalars['Boolean'];
};

export type Pick = {
  __typename?: 'Pick';
  game?: Maybe<Game>;
  id: Scalars['ID'];
  isCorrect?: Maybe<Scalars['Boolean']>;
  picked?: Maybe<Team>;
  player?: Maybe<Player>;
};
// ...
export type Player = {
  __typename?: 'Player';
  email?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  name?: Maybe<Scalars['String']>;
  password?: Maybe<PasswordState>;
  passwordResetIssuedAt?: Maybe<Scalars['String']>;
  passwordResetRedeemedAt?: Maybe<Scalars['String']>;
  passwordResetToken?: Maybe<PasswordState>;
  picks?: Maybe<Array<Pick>>;
  picksCount?: Maybe<Scalars['Int']>;
};
// ...
export type Team = {
  __typename?: 'Team';
  abbreviation?: Maybe<Scalars['String']>;
  city?: Maybe<Scalars['String']>;
  conference?: Maybe<Scalars['String']>;
  division?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  name?: Maybe<Scalars['String']>;
};